+++ R/ggsurvtheme.R
@@ -1,7 +1,7 @@
 #' @include utilities.R
 NULL
 
-#' Default Theme for Survminer Plots
+#' Theme for Survminer Plots
 #'
 #' @description
 #' Default theme for plots generated with survminer.
@@ -17,7 +17,7 @@ NULL
 #'  c("top", "bottom", "left", "right", "none"). Default is "top" side position.
 #'  to remove the legend use legend = "none". Legend position can be also
 #'  specified using a numeric vector c(x, y); see details section.
-#' @name theme_survminer
+#' @param ... additional arguments passed to the function theme_survminer().
 #' @examples
 #'
 #'# Fit survival curves
@@ -50,6 +50,9 @@ NULL
 #'      font.tickslab = c(12, "plain", "darkgreen")
 #'    )
 #'  )
+#'
+#'
+#' @describeIn ggsurvtheme Default theme for survminer plots. A theme similar to theme_classic() with large font size.
 #' @export
 theme_survminer <-
   function (base_size = 12, base_family = "",
@@ -58,7 +61,8 @@ theme_survminer <-
             font.caption = c(15, "plain", "black"),
             font.tickslab = c(12, "plain", "black"),
             legend = c("top", "bottom", "left", "right", "none"),
-            font.legend = c(10, "plain", "black")
+            font.legend = c(10, "plain", "black"),
+            ...
             )
   {
 
@@ -119,3 +123,20 @@ theme_survminer <-
     result
   }
 
+
+#' @export
+#' @describeIn ggsurvtheme theme for drawing a clean risk table and cumulative
+#'   number of events table. A theme similar to theme_survminer() without i)
+#'   axis lines and, ii) x axis ticks and title.
+theme_cleantable <- function(base_size = 12, base_family = "", ...)
+{
+  theme_survminer(base_size = base_size, base_family = base_family, ...) %+replace%
+    theme(
+      axis.line.x = element_blank(),
+      axis.line.y = element_blank(),
+      axis.text.x = element_blank(),
+      axis.ticks.x = element_blank(),
+      axis.title.x = element_blank(),
+      axis.title.y = element_blank()
+    )
+}